% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcfa-app.R
\name{mgcfa_app}
\alias{mgcfa_app}
\title{Build the Interactive AutomaticMGCFA App}
\usage{
mgcfa_app()
}
\value{
A \code{shiny.appobj}.
}
\description{
Creates and returns the Shiny app object used by
\code{mgcfa_launch_app()}.
}
