% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcfa.R
\name{mgcfa_help_inputs}
\alias{mgcfa_help_inputs}
\title{Quick Input Guidance for AutomaticMGCFA}
\usage{
mgcfa_help_inputs()
}
\value{
Invisibly returns a data.frame of recipes.
}
\description{
Prints concise recipes for common input types.
}
