% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcfa.R
\name{print.mgcfa_bias_effects}
\alias{print.mgcfa_bias_effects}
\title{Print Method for Bias Effect Summaries}
\usage{
\method{print}{mgcfa_bias_effects}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{mgcfa_bias_effects()}.}

\item{...}{Unused.}
}
\value{
The input object invisibly.
}
\description{
Print Method for Bias Effect Summaries
}
