% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcfa.R
\name{print.mgcfa_input}
\alias{print.mgcfa_input}
\title{Print Method for Prepared MGCFA Inputs}
\usage{
\method{print}{mgcfa_input}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{mgcfa_prepare_input()}.}

\item{...}{Unused.}
}
\value{
The input object invisibly.
}
\description{
Print Method for Prepared MGCFA Inputs
}
