% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcfa-bias.R
\name{print.mgcfa_report}
\alias{print.mgcfa_report}
\title{Print Method for MGCFA Reports}
\usage{
\method{print}{mgcfa_report}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{mgcfa_report()}.}

\item{...}{Unused.}
}
\value{
The input object invisibly.
}
\description{
Print Method for MGCFA Reports
}
